refactor: pull most of main out into an app-specific update function
